requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
openai>=1.0.0
python-dotenv
